requirements.txt
apache-airflow==2.0.1
SQLAlchemy==1.3.23
pandas
requests
openpyxl==3.0.6
psycopg2==2.9.5
psycopg2-binary==2.8.6
urllib3==1.26.9
pyarrow==3.0.0
supervisor==4.2.2
pathlib==1.0.1
wtforms==2.3.3
pendulum
xlrd
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1